+++ theme/templates/base.html
@@ -14,8 +14,8 @@
     <header>
         <h1 class="text-3xl" >Gear List</h1>
         <nav>
-            <a href="{% url 'snack_list' %}">My Gear List</a>
-            <a href="{% url 'snack_create' %}">Add a new item</a>
+            <a href="{% url 'gear_list' %}">My Gear List</a>
+            <a href="{% url 'gear_create' %}">Add a new item</a>
             <a href="{% url 'about' %}">About Gear List</a>
 
         </nav>
