+++ theme/templates/gear_delete.html
@@ -1,0 +1,13 @@
+{% extends 'base.html' %}
+
+{% block crud %}
+
+<h1>Delete Gear Item</h1>
+
+<form action="" method="POST">
+  {% csrf_token %}
+  <p>Do you really want to delete "{{ gear.name }}"?</p>
+  <input type="submit" value="OK">
+</form>
+
+{% endblock crud %}
