--- theme/templates/snack_delete.html
@@ -1,13 +1,0 @@
-{% extends 'base.html' %}
-
-{% block crud %}
-
-<h1>Delete Gear Item</h1>
-
-<form action="" method="POST">
-  {% csrf_token %}
-  <p>Do you really want to delete "{{ snack.name }}"?</p>
-  <input type="submit" value="OK">
-</form>
-
-{% endblock crud %}
