--- theme/templates/snack_detail.html
@@ -1,11 +1,0 @@
-{% extends 'base.html' %}
-
-{% block crud %}
-
-<h2>{{ snack.name }}</h2>
-<h3>Backpacker: {{ snack.purchaser }} </h3>
-<p>Gear description: {{ snack.description }}</p>
-
-<a href="{% url 'snack_update' snack.pk %}">+ Update Gear Item</a>
-<a href="{% url 'snack_delete' snack.pk %}">+ Delete Gear Item</a>
-{% endblock crud %}
