--- theme/templates/snack_list.html
@@ -1,11 +1,0 @@
-{% extends 'base.html' %}
-
-{% block crud %}
-<ul>
-    {% for snack in object_list %}
-        <li>
-            <a href="{% url 'snack_detail' snack.pk %}">{{ snack.name }}</a>
-        </li>
-    {% endfor %}
-</ul>
-{% endblock crud %}
